LayoutTests/fast/selectors/empty-namespace-with-element-selector-expected.txt

Verify that "|name" matches the empty namespace while "*|name" and "name" match any namespace.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Trivial Selector used as rightmost
PASS root.querySelectorAll("|div").length is 1
PASS root.querySelectorAll("|div")[0].getAttribute("data-case") is "2"
PASS root.querySelectorAll("*|div").length is 2
PASS root.querySelectorAll("*|div")[0].getAttribute("data-case") is "1"
PASS root.querySelectorAll("*|div")[1].getAttribute("data-case") is "2"
PASS root.querySelectorAll("div").length is 2
PASS root.querySelectorAll("div")[0].getAttribute("data-case") is "1"
PASS root.querySelectorAll("div")[1].getAttribute("data-case") is "2"
PASS getComputedStyle(document.querySelector("[|data-case='1' i]")).color is "rgb(4, 5, 6)"
PASS getComputedStyle(document.querySelector("[|data-case='1' i]")).backgroundColor is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelector("[|data-case='2' i]")).color is "rgb(1, 2, 3)"
PASS getComputedStyle(document.querySelector("[|data-case='2' i]")).backgroundColor is "rgb(4, 5, 6)"
PASS successfullyParsed is true

TEST COMPLETE
